fix(organisation): organize code in a simple way
Issue: #23
