+++ requirements.txt
@@ -2,8 +2,8 @@ attrs==19.3.0
 aws-sam-translator==1.19.0
 aws-xray-sdk==2.4.3
 boto==2.49.0
-boto3==1.10.39
-botocore==1.13.39
+boto3==1.10.40
+botocore==1.13.40
 certifi==2019.11.28
 cffi==1.13.2
 cfn-lint==0.26.0
@@ -17,7 +17,7 @@ idna==2.8
 importlib-metadata==1.3.0
 Jinja2==2.10.3
 jmespath==0.9.4
-jsondiff==1.1.2
+jsondiff==1.2.0
 jsonpatch==1.24
 jsonpickle==1.2
 jsonpointer==2.0
@@ -31,12 +31,12 @@ pluggy==0.13.1
 py==1.8.0
 pyasn1==0.4.8
 pycparser==2.19
-pynamodb==4.0.0
+pynamodb==4.2.0
 pyparsing==2.4.5
 pyrsistent==0.15.6
-pytest==5.3.1
-python-dateutil==2.8.0
+pytest==5.3.2
 pytest-mock==1.13.0
+python-dateutil==2.8.1
 python-jose==3.1.0
 pytz==2019.3
 PyYAML==5.2
@@ -46,7 +46,7 @@ rsa==4.0
 s3transfer==0.2.1
 six==1.13.0
 sshpubkeys==3.1.0
-urllib3==1.25.6
+urllib3==1.25.7
 wcwidth==0.1.7
 websocket-client==0.56.0
 Werkzeug==0.16.0
