+++ template.yaml
@@ -19,7 +19,7 @@ Resources:
   JobbieStoreFunction:
     Type: AWS::Serverless::Function
     Properties:
-      CodeUri: app/
+      CodeUri: .
       Handler: app.lambda_handler
       Runtime: python3.7
       Events:
